// src/server/fetch-clients/paystack/schemas.ts
export const baseResponseSchema = z.object({
  status: z.boolean(),
  message: z.string(),
});
// ...
export const createSubaccountSchema = z.object({
  business_name: z.string().describe("Name of business for subaccount"),
  settlement_bank: z.string().describe("Bank code for the bank"),
  account_number: z.string().describe("Bank account number"),
  percentage_charge: z
    .number()
    .default(SUBACCOUNT_PERCENTAGE_CHARGE)
    .describe(
      "The default percentage charged when receiving on behalf of this subaccount",
    ),
  description: z
    .string()
    .optional()
    .describe("A description for this subaccount"),
  primary_contact_email: z
    .email()
    .optional()
    .describe("A contact email for the subaccount"),
  primary_contact_name: z
    .string()
    .optional()
    .describe("A name for the contact person for this subaccount"),
  primary_contact_phone: z
    .string()
    .optional()
    .describe("A phone number to call for this subaccount"),
  metadata: z
    .record(z.string(), z.unknown())
    .optional()
    .describe("Stringified JSON object of custom data"),
});

export const subaccountSchema = z.object({
  id: z.number(),
  subaccount_code: z.string(),
  business_name: z.string(),
  description: z.string().optional(),
  primary_contact_name: z.string().nullable().optional(),
  primary_contact_email: z.string().nullable().optional(),
  primary_contact_phone: z.string().nullable().optional(),
  metadata: z.record(z.string(), z.unknown()).nullable().optional(),
  percentage_charge: z.number(),
  settlement_bank: z.string(),
  account_number: z.string(),
  settlement_schedule: z
    .string()
    // .enum(["auto", "weekly", "monthly", "manual"])
    .optional(),
  active: z.boolean(),
  migrate: z.boolean().optional(),
  createdAt: z.string(),
  updatedAt: z.string(),
});